Rework library to be allocation free
Now the api is simpler, and introduces two top level functions:

- `from_str`
- `from`

The point is that we either take a string representation of a UUID, or just the
UUID as a `[16]u8`, and as such we know the size of the id to create. No need to
allocate, just return the array as is.

Coerce to a slice as per usual with `&`.